Recommender System/requirements.txt
pandas==1.3.2
pymysql==1.0.2
konlpy==0.5.2
flask
python-dotenv